+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Notes for Craft Commerce
 
+## Unreleased
+
+- Fixed a bug where empty variant collections could be incorrectly memoized when duplicating a product. ([#4075](https://github.com/craftcms/commerce/issues/4075))
+
 ## 5.4.4 - 2025-07-31
 
 - Improved the performance of loading Edit Order pages. ([#4056](https://github.com/craftcms/commerce/issues/4056))